watchdog/requirements.txt
uvicorn==0.27.1
fastapi==0.110.0
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.27.0
PyYAML==6.0.1 